Refactor: verifyTicketOwner - admin package 로 이동
