+++ src/main/java/ticketaka/mtvs3_final_backend/admin/command/application/dto/AdminVerificationResponseDTO.java
@@ -1,5 +1,7 @@
 package ticketaka.mtvs3_final_backend.admin.command.application.dto;
 
+import org.springframework.web.multipart.MultipartFile;
+
 import java.time.LocalDateTime;
 
 public class AdminVerificationResponseDTO {
@@ -9,6 +11,13 @@ public class AdminVerificationResponseDTO {
             String concertName,
             LocalDateTime concertDate,
             String seatInfo
+    ) {
+    }
+
+    // 티켓 사용 파일 업로드
+    public record verifyTicketOwnerDTO(
+            MultipartFile image,
+            Long ticketId
     ) {
     }
 }
